quality-of-life fixes (#178)
* cleanup flush_tx_cache

- remove code from unnecessary else branch
- enforce minimum 2-block wait period before removing stalled txs

* make fill-block get closer to block limit

* change default interval & num_txs for fill-block, change some usizes to u64s

* clippy: remove unnecessary cast

+++ crates/core/src/db/mod.rs
@@ -48,7 +48,7 @@ pub trait DbOps {
     fn create_tables(&self) -> Result<()>;
 
     /// Insert a new run into the database. Returns run_id.
-    fn insert_run(&self, timestamp: u64, tx_count: usize, scenario_name: &str) -> Result<u64>;
+    fn insert_run(&self, timestamp: u64, tx_count: u64, scenario_name: &str) -> Result<u64>;
 
     fn version(&self) -> u64;
